requirements.txt
Django==1.8.2
argparse==1.2.1
djangorestframework==3.1.3
ecdsa==0.13
mongoengine==0.10.0
paramiko==1.15.2
pycrypto==2.6.1
pymongo==3.0.2
wsgiref==0.1.2
